# tsuba: move parquet transformations into async tasks