--- a/libtsuba/CMakeLists.txt
+++ b/libtsuba/CMakeLists.txt
@@ -14,6 +14,7 @@ add_dependencies(lib tsuba tsuba-preload)
 
 set(sources
   src/AddProperties.cpp
+  src/AsyncOpGroup.cpp
   src/Errors.cpp
   src/FaultTest.cpp
   src/file.cpp
@@ -34,6 +35,7 @@ set(sources
   src/RDGPartHeader.cpp
   src/RDGPrefix.cpp
   src/RDGSlice.cpp
+  src/ReadGroup.cpp
   src/tsuba.cpp
   src/WriteGroup.cpp
 )