--- a/libtsuba/src/AddProperties.cpp
+++ b/libtsuba/src/AddProperties.cpp
@@ -1,5 +1,7 @@
 #include "AddProperties.h"
 
+#include <memory>
+
 #include <arrow/chunked_array.h>
 
 #include "katana/Result.h"
@@ -52,7 +54,7 @@ tsuba::LoadProperties(
     return DoLoadProperties(expected_name, file_path);
   } catch (const std::exception& exp) {
     return KATANA_ERROR(
-        ErrorCode::ArrowError, "arrow exception: {}", exp.what());
+        tsuba::ErrorCode::ArrowError, "arrow exception: {}", exp.what());
   }
 }
 
@@ -63,9 +65,107 @@ tsuba::LoadPropertySlice(
   try {
     return DoLoadProperties(
         expected_name, file_path,
-        ParquetReader::Slice{.offset = offset, .length = length});
+        tsuba::ParquetReader::Slice{.offset = offset, .length = length});
   } catch (const std::exception& exp) {
     return KATANA_ERROR(
-        ErrorCode::ArrowError, "arrow exception: {}", exp.what());
+        tsuba::ErrorCode::ArrowError, "arrow exception: {}", exp.what());
   }
+}
+
+katana::Result<void>
+tsuba::AddProperties(
+    const katana::Uri& uri,
+    const std::vector<tsuba::PropStorageInfo>& properties, ReadGroup* grp,
+    const std::function<katana::Result<void>(std::shared_ptr<arrow::Table>)>&
+        add_fn) {
+  for (const tsuba::PropStorageInfo& prop : properties) {
+    const std::string& name = prop.name;
+    const katana::Uri& path = uri.Join(prop.path);
+    std::future<katana::Result<std::shared_ptr<arrow::Table>>> future =
+        std::async(
+            std::launch::async,
+            [name, path]() -> katana::Result<std::shared_ptr<arrow::Table>> {
+              auto load_result = LoadProperties(name, path);
+              if (!load_result) {
+                return load_result.error().WithContext(
+                    "error loading {}", path);
+              }
+              return load_result.value();
+            });
+    auto on_complete = [add_fn,
+                        name](const std::shared_ptr<arrow::Table>& props)
+        -> katana::Result<void> {
+      auto add_result = add_fn(props);
+      if (!add_result) {
+        return add_result.error().WithContext("adding {}", std::quoted(name));
+      }
+      return katana::ResultSuccess();
+    };
+    if (grp) {
+      grp->AddReturnsOp<std::shared_ptr<arrow::Table>>(
+          std::move(future), path.string(), on_complete);
+      continue;
+    }
+    auto read_res = future.get();
+    if (!read_res) {
+      return read_res.error();
+    }
+    auto on_complete_res = on_complete(read_res.value());
+    if (!on_complete_res) {
+      return on_complete_res.error();
+    }
+  }
+
+  return katana::ResultSuccess();
+}
+
+katana::Result<void>
+tsuba::AddPropertySlice(
+    const katana::Uri& dir,
+    const std::vector<tsuba::PropStorageInfo>& properties,
+    std::pair<uint64_t, uint64_t> range, ReadGroup* grp,
+    const std::function<katana::Result<void>(std::shared_ptr<arrow::Table>)>&
+        add_fn) {
+  uint64_t begin = range.first;
+  uint64_t size = range.second - range.first;
+  for (const tsuba::PropStorageInfo& prop : properties) {
+    const std::string& name = prop.name;
+    const katana::Uri& path = dir.Join(prop.path);
+    std::future<katana::Result<std::shared_ptr<arrow::Table>>> future =
+        std::async(
+            std::launch::async,
+            [name, path, begin,
+             size]() -> katana::Result<std::shared_ptr<arrow::Table>> {
+              auto load_result = LoadPropertySlice(name, path, begin, size);
+              if (!load_result) {
+                return load_result.error().WithContext(
+                    "error loading {}", path);
+              }
+              return load_result.value();
+            });
+    auto on_complete = [add_fn,
+                        name](const std::shared_ptr<arrow::Table>& props)
+        -> katana::Result<void> {
+      auto add_result = add_fn(props);
+      if (!add_result) {
+        return add_result.error().WithContext("adding {}", std::quoted(name));
+      }
+      return katana::ResultSuccess();
+    };
+    if (grp) {
+      grp->AddReturnsOp<std::shared_ptr<arrow::Table>>(
+          std::move(future), path.string(), on_complete);
+      continue;
+    }
+    auto read_res = future.get();
+    if (!read_res) {
+      return read_res.error();
+    }
+    auto on_complete_res = on_complete(read_res.value());
+    if (!on_complete_res) {
+      return on_complete_res.error();
+    }
+  }
+
+  return katana::ResultSuccess();
 }