--- a/libtsuba/src/AddProperties.h
+++ b/libtsuba/src/AddProperties.h
@@ -6,6 +6,7 @@
 #include "RDGPartHeader.h"
 #include "katana/Result.h"
 #include "katana/Uri.h"
+#include "tsuba/ReadGroup.h"
 
 namespace tsuba {
 
@@ -16,56 +17,18 @@ KATANA_EXPORT katana::Result<std::shared_ptr<arrow::Table>> LoadPropertySlice(
     const std::string& expected_name, const katana::Uri& file_path,
     int64_t offset, int64_t length);
 
-template <typename AddFn>
-katana::Result<void>
-AddProperties(
+KATANA_EXPORT katana::Result<void> AddProperties(
     const katana::Uri& uri,
-    const std::vector<tsuba::PropStorageInfo>& properties, AddFn add_fn) {
-  for (const tsuba::PropStorageInfo& properties : properties) {
-    auto p_path = uri.Join(properties.path);
+    const std::vector<tsuba::PropStorageInfo>& properties, ReadGroup* grp,
+    const std::function<katana::Result<void>(std::shared_ptr<arrow::Table>)>&
+        add_fn);
 
-    auto load_result = LoadProperties(properties.name, p_path);
-    if (!load_result) {
-      return load_result.error().WithContext("error loading {}", p_path);
-    }
-
-    std::shared_ptr<arrow::Table> props = load_result.value();
-
-    auto add_result = add_fn(props);
-    if (!add_result) {
-      return add_result.error().WithContext(
-          "adding {}", std::quoted(properties.name));
-    }
-  }
-
-  return katana::ResultSuccess();
-}
-
-template <typename AddFn>
-katana::Result<void>
-AddPropertySlice(
+KATANA_EXPORT katana::Result<void> AddPropertySlice(
     const katana::Uri& dir,
     const std::vector<tsuba::PropStorageInfo>& properties,
-    std::pair<uint64_t, uint64_t> range, AddFn add_fn) {
-  for (const tsuba::PropStorageInfo& properties : properties) {
-    katana::Uri p_path = dir.Join(properties.path);
-
-    auto load_result = LoadPropertySlice(
-        properties.name, p_path, range.first, range.second - range.first);
-    if (!load_result) {
-      return load_result.error();
-    }
-
-    std::shared_ptr<arrow::Table> props = load_result.value();
-
-    auto add_result = add_fn(props);
-    if (!add_result) {
-      return add_result.error();
-    }
-  }
-
-  return katana::ResultSuccess();
-}
+    std::pair<uint64_t, uint64_t> range, ReadGroup* grp,
+    const std::function<katana::Result<void>(std::shared_ptr<arrow::Table>)>&
+        add_fn);
 
 }  // namespace tsuba