--- a/libtsuba/src/ReadGroup.cpp
+++ b/libtsuba/src/ReadGroup.cpp
@@ -1,0 +1,13 @@
+#include "tsuba/ReadGroup.h"
+
+void
+tsuba::ReadGroup::AddOp(
+    std::future<katana::Result<void>> future, std::string file,
+    const std::function<katana::Result<void>()>& on_complete) {
+  async_op_group_.AddOp(std::move(future), std::move(file), on_complete);
+}
+
+katana::Result<void>
+tsuba::ReadGroup::Finish() {
+  return async_op_group_.Finish();
+}