… libtsuba/src/WriteGroup.cpp