@tylershunt 2021-04-26
Updated on the read side as well. Will put in an enterprise PR that points to this to verify that it still passes all of our tests.